+++ apx_sdk/CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Implemented `PartialEq` on `apx_core::ap_url::ApUrl`.
 - Implemented `PartialEq` on `apx_sdk::url::Url`.
+- Added support for fragment resolution to `fetch_object`.
 
 ### Changed